\name{REORS-package}
\alias{REORS-package}
\alias{REORS}
\docType{package}

\title{
Package for remote sensing and Earth observation in R.
}

\description{
While the raster package includes many tools to work with RS data, the actual implementation is left up to the end-user.
This package is intended to help fill that gap, by implementing functionality present in commercial software and algorithms in scientific papers.
}

\details{
\tabular{ll}{
Package: \tab REORS\cr
Type: \tab Package\cr
Version: \tab 0.1-7\cr
Date: \tab 2014-08-02\cr
License: \tab GPL3\cr
}

}
\author{
Joseph McGrath <joseph.mcgrath@outlook.com>
}

\references{
Listed with individual function.
}

\keyword{ package }

\examples{
~~ simple examples of the most important functions ~~
}
